requirements.txt
# GPT4 PineScript V5 Strategy Generator
# Python 3.11+

openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
